refactor(guild): move GuildError to dedicated error.rs
